modularize images in post
