+++ Mobile/starter_project_matador/lib/features/profile_page/presentation/widget/status_card.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/src/widgets/framework.dart';
-import 'package:flutter/src/widgets/placeholder.dart';
+
 
 class StatusCard extends StatelessWidget {
   const StatusCard({super.key});
